Fix generated types
