+++ react/IOMessage.tsx
@@ -1,8 +1,13 @@
-import React, { Fragment } from 'react'
-import { useIntl } from 'react-intl'
+import React, { Fragment, FC } from 'react'
+import { useIntl, MessageDescriptor } from 'react-intl'
 
-import formatIOMessage from './formatIOMessage'
-import { IOMessage as IOMessageType } from './types/IOMessage'
+import formatIOMessage, { Values } from './formatIOMessage'
+
+type IOMessageType = FC<
+  MessageDescriptor & {
+    values?: Values
+  }
+>
 
 const IOMessage: IOMessageType = ({
   children,
