--- react/types/IOMessage.ts
@@ -1,10 +1,0 @@
-import { FunctionComponent } from 'react'
-import { MessageDescriptor } from 'react-intl'
-
-import { Values } from './formatIOMessage'
-
-export type IOMessage = FunctionComponent<
-  MessageDescriptor & {
-    values?: Values
-  }
->
